LayoutTests/fast/custom-elements/Document-defineCustomElement.html

<!DOCTYPE html>
<html>
<head>
<title>Custom Elements: Extensions to Document interface</title>
<meta name="author" title="Ryosuke Niwa" href="mailto:rniwa@webkit.org">
<meta name="assert" content="document.defineCustomElement should define a custom element">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<link rel='stylesheet' href='../../resources/testharness.css'>
</head>
<body>
<div id="log"></div>
<script>

test(function () {
    assert_true('defineCustomElement' in Document.prototype, '"defineCustomElement" exists on Document.prototype');
    assert_true('defineCustomElement' in document, '"defineCustomElement" exists on document');
}, 'Check the existence of defineCustomElement on Document interface');

test(function () {
    class MyCustomElement extends HTMLElement {};

    assert_throws({'name': 'SyntaxError'}, function () { document.defineCustomElement(null, MyCustomElement); },
        'document.defineCustomElement must throw a SyntaxError if the tag name is null');
    assert_throws({'name': 'SyntaxError'}, function () { document.defineCustomElement('', MyCustomElement); },
        'document.defineCustomElement must throw a SyntaxError if the tag name is empty');
    assert_throws({'name': 'SyntaxError'}, function () { document.defineCustomElement('abc', MyCustomElement); },
        'document.defineCustomElement must throw a SyntaxError if the tag name does not contain "-"');
    assert_throws({'name': 'SyntaxError'}, function () { document.defineCustomElement('a-Bc', MyCustomElement); },
        'document.defineCustomElement must throw a SyntaxError if the tag name contains an upper case letter');

    var builtinTagNames = [
        'annotation-xml',
        'color-profile',
        'font-face',
        'font-face-src',
        'font-face-uri',
        'font-face-format',
        'font-face-name',
        'missing-glyph'
    ];

    for (var tagName of builtinTagNames) {
        assert_throws({'name': 'SyntaxError'}, function () { document.defineCustomElement(tagName, MyCustomElement); },
            'document.defineCustomElement must throw a SyntaxError if the tag name is "' + tagName + '"');
    }

}, 'document.defineCustomElement should throw with an invalid name');

test(function () {
    class SomeCustomElement extends HTMLElement {};
    class OtherCustomElement extends HTMLElement {};

    document.defineCustomElement('some-custom-element', SomeCustomElement);
    assert_throws({'name': 'NotSupportedError'}, function () { document.defineCustomElement('some-custom-element', OtherCustomElement); },
        'document.defineCustomElement must throw a NotSupportedError if the specified tag name is already used');

}, 'document.defineCustomElement should throw with a duplicate name');

test(function () {
    assert_throws({'name': 'TypeError'}, function () { document.defineCustomElement('invalid-element', 1); },
        'document.defineCustomElement must throw a TypeError when the element interface is a number');
    assert_throws({'name': 'TypeError'}, function () { document.defineCustomElement('invalid-element', '123'); },
        'document.defineCustomElement must throw a TypeError when the element interface is a string');
    assert_throws({'name': 'TypeError'}, function () { document.defineCustomElement('invalid-element', {}); },
        'document.defineCustomElement must throw a TypeError when the element interface is an object');
    assert_throws({'name': 'TypeError'}, function () { document.defineCustomElement('invalid-element', []); },
        'document.defineCustomElement must throw a TypeError when the element interface is an array');
}, 'document.defineCustomElement should throw when the element interface is not a constructor');

test(function () {
    class MyCustomElement extends HTMLElement {};
    document.defineCustomElement('my-custom-element', MyCustomElement);

    var instance = new MyCustomElement;
    assert_true(instance instanceof MyCustomElement,
        'An instance of a custom HTML element be an instance of the associated interface');

    assert_true(instance instanceof HTMLElement,
        'An instance of a custom HTML element must inherit from HTMLElement');

    assert_equals(instance.localName, 'my-custom-element',
        'An instance of a custom element must use the associated tag name');

    assert_equals(instance.namespaceURI, 'http://www.w3.org/1999/xhtml',
        'A custom element HTML must use HTML namespace');

}, 'document.defineCustomElement should define an instantiatable custom element');

</script>
</body>
</html>
