LayoutTests/fast/custom-elements/Document-defineCustomElement-expected.txt


PASS Check the existence of defineCustomElement on Document interface 
PASS document.defineCustomElement should throw with an invalid name 
PASS document.defineCustomElement should throw with a duplicate name 
PASS document.defineCustomElement must throw a NotSupportedError when the context object is an associated inert template document 
PASS document.defineCustomElement must throw a NotSupportedError when the context object is created by DOMImplementation.createHTMLDocument 
PASS document.defineCustomElement must throw a NotSupportedError when the context object is created by DOMImplementation.createDocument 
PASS document.defineCustomElement should throw when the element interface is not a constructor 
PASS document.defineCustomElement should define an instantiatable custom element 
